+++ packages/scripts/CHANGELOG.md
@@ -5,6 +5,8 @@ Unreleased
 ----------
 
 + Added a blocks-specific webpack config: `./config/blocks-webpack.config.js`.
++ Removes the webpack copy plugin responsible for copying `block.json` files from the `src/` directory to the `${outputDir}/blocks` directory. 
+
 
 v3.2.0 - 2022-07-12
 -------------------